Practice/requirements.txt
pytorch=1.9.0
numpy
torchvision
tensorflow
